feat: add help command to display usage information for mcpr CLI
