+++ packages/cli/package.json
@@ -2,12 +2,12 @@
   "name": "@mcp_router/cli",
   "version": "0.1.3",
   "description": "CLI for MCP Router - connect to an MCP HTTP server",
-  "main": "dist/mcpr.js",
+  "main": "dist/cli/mcpr.js",
   "engines": {
     "node": ">=18.0.0"
   },
   "bin": {
-    "mcpr-cli": "dist/mcpr.js"
+    "mcpr-cli": "dist/cli/mcpr.js"
   },
   "type": "module",
   "scripts": {
