+++ packages/cli/src/commands/serve.ts
@@ -2,7 +2,7 @@ import { createServer } from "http";
 import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
 import { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
-import { MCPAggregator } from "../../mcp-aggregator.js";
+import { MCPAggregator } from "../mcp-aggregator.js";
 
 /**
  * Executes the serve command, starting an HTTP server that accepts
