+++ packages/cli/src/commands/version.ts
